Add files via upload
The .cpp file is the backend for the app, and is also a web server.  It also declares two environment variables that the application requires: the API Key for Google Maps, and the Access Key for the currencylayer.com currency API.  Without the former, the application won't be visible because it's a Google Maps app with a currency conversion form appearing on the info window.  And without the latter the app won't work correctly because it needs that key to be able to get a list of currencies as well as to get the conversion rates.  Of course, the .exe file is the (Windows) compiled executable for the backend app.